Move get_tftest_directive to a separate file
